Add physics constraint modules for simulation-free nonlocal training
Implement the constraint losses, correlation functions, and validation
infrastructure needed to train the nonlocal neural functional without
simulation data — using only exact thermodynamic sum rules and
consistency conditions.

New modules:
- constraints/ — contact sum rule, compressibility sum rule, Gibbs
  adsorption, pressure/OZ consistency, Noether translational
  invariance, scaled-particle theory limits, positivity
- correlations/ — c₂ via JAX autodiff (forward-over-reverse for O(N)),
  Ornstein-Zernike solver in Fourier space, structure factor S(k),
  pair correlation g(r)
- validation/ — analytical Percus-Yevick c₂, S(k), g(r) reference
  solutions for hard spheres

Extended training/losses.py with nonlocal bulk loss, constraint loss,
and combined loss functions. Added constraint weight fields to
TrainingConfig. Fixed JAX tracing bug in bulk_parameters (float→jnp).

+++ nonlocal_ext/functional.py
@@ -259,7 +259,7 @@ class NonlocalLutskoFunctional(eqx.Module):
         A, B = self.predict_parameters(rho_uniform)
 
         # In uniform system, A and B should be constant
-        return float(jnp.mean(A)), float(jnp.mean(B))
+        return jnp.mean(A), jnp.mean(B)
 
     def __repr__(self) -> str:
         return (f"NonlocalLutskoFunctional(\n"
